fix: rename data.ts to _data.ts to prevent Gatsby page creation
