+++ src/pages/resources/_data.ts
